+++ sdk/core/azure-core/azure/core/pipeline/transport/_base.py
@@ -112,6 +112,18 @@ def _case_insensitive_dict(*args, **kwargs):
 
 
 def _format_url_section(template, **kwargs):
+    """String format the template with the kwargs, auto-skip sections of the template that are NOT in the kwargs.
+
+    By default in Python, "format" will raise a KeyError if a template element is not found. Here the section between
+    the slashes will be removed from the template instead.
+
+    This is used for API like Storage, where when Swagger has template section not defined as parameter.
+
+    :param str template: a string template to fill
+    :param dict[str,str] kwargs: Template values as string
+    :rtype: str
+    :returns: Template completed
+    """
     components = template.split("/")
     while components:
         try:
@@ -718,7 +730,12 @@ class PipelineClientBase(object):
             parsed = urlparse(url)
             if not parsed.scheme or not parsed.netloc:
                 url = url.lstrip("/")
-                base = self._base_url.format(**kwargs).rstrip("/")
+                try:
+                    base = self._base_url.format(**kwargs).rstrip("/")
+                except KeyError as key:
+                    err_msg = "The value provided for the url part {} was incorrect, and resulted in an invalid url"
+                    raise ValueError(err_msg.format(key.args[0]))
+
                 url = _urljoin(base, url)
         else:
             url = self._base_url.format(**kwargs)