+++ sdk/identity/azure-identity/tests/test_vscode_credential_async.py
@@ -106,17 +106,3 @@ async def test_no_obtain_token_if_cached():
         credential = VSCodeCredential(_client=mock_client)
         token = await credential.get_token("scope")
         assert token_by_refresh_token.call_count == 0
-
-
-@pytest.mark.skipif(not sys.platform.startswith("linux"), reason="This test only runs on Linux")
-@pytest.mark.asyncio
-async def test_distro():
-
-    mock_client = mock.Mock(spec=object)
-    token_by_refresh_token = mock.Mock(return_value=None)
-    mock_client.obtain_token_by_refresh_token = wrap_in_future(token_by_refresh_token)
-    mock_client.get_cached_access_token = mock.Mock(return_value=None)
-
-    with pytest.raises(CredentialUnavailableError):
-        credential = VSCodeCredential(_client=mock_client)
-        token = await credential.get_token("scope")