+++ sdk/core/azure-core/CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Bug fixes
 
 - `AzureKeyCredentialPolicy` will now accept (and ignore) passed in kwargs  #11963
+- Better error messages if passed endpoint is incorrect  #12106
 
 ## 1.6.0 (2020-06-03)